Learn Python Quickly Source Code_v2/code3/pgvector-python-starter/requirements.txt

numpy==1.24.4
scipy==1.10.1
torch==2.1.0
sentence-transformers
python-dotenv
psycopg2-binary

